### 2020-09-24 #400 @ClaudioZandonella
Presubmission inquiry: PRDA
Submitting Author: Claudio Zandonella Callegher (@ClaudioZandonella)
Repository: https://github.com/ClaudioZandonella/PRDA

---

```
Package: PRDA
Title: Conduct a Prospective or Retrospective Design Analysis
Version: 1.0.0.9000
    person("Claudio", "Zandonella Callegher", role = c("aut", "cre"),
           email = "claudiozandonella@gmail.com",
    person("Massimiliano", "Pastore", role = c("aut"),
           email = "massimiliano.pastore@unipd.it",
    person("Angela", "Andreella", role = c("aut"),
           email = "a.andreella8@gmail.com",
    person("Anna", "Vesely", role = c("aut"),
           email = "anna.vesely92@gmail.com",
    person("Enrico", "Toffalini", role = c("aut"),
           email = "enrico.toffalini@yahoo.it",
    person("Giulia", "Bertoldo", role = c("aut"),
           email = "giuliabertoldo94@gmail.com",
    person("Gianmarco", "Altoè", role = c("aut"),
           email = "gianmarco.altoe@unipd.it",
           comment = c(ORCID = "0000-0003-1154-9528")))
Description: An implementation of the "Design Analysis" proposed by
    Gelman and Carlin (2014). It combines the evaluation of Power-Analysis
    with other inferential-risks as Type-M error (e.g. Magnitude) and Type-S
    error (e.g. Sign). See also Altoè et al. (2020) and Bertoldo et al. (2020).
    covr,
    devtools,
    roxygen2,
    tidyverse
    Rcpp, RcppArmadillo
    MASS,
    Rcpp
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

The pakage allows conducting a Design Analysis to evaluate power levels together with estimates uncertainty (i.e., Type M error and Type S error) under hypothetical replications of a study. This is useful in the planning stage of a study to define the required sample size, or to evaluate the inferential risks associated with a study that has been already conducted.

Researchers that want to preregister their studies and need to conduct a power analysis, or researcher that want to evaluate inferential risks related to their studies when interpreting the results. In particular, researchers in Psychology or Social Sciences.

Another recent R package, `retrodesign` (https://github.com/csetraynor/retrodesign), allows conducting retrospective design analysis considering the estimate of the unstandardized effect size (i.e., regression coefficient or mean difference) and standard error of the estimate. `PRDA` package, instead, considers standardized effect size (i.e., Pearson correlation coefficient or Cohen’s d) and study sample size. These are more commonly used in research fields such as Psychology or Social Science, and therefore are implemented in PRDA to facilitate researchers’ reasoning about design analysis.  PRDA, additionally, offers the possibility to conduct a prospective design analysis and to account for the uncertainty about the hypothetical value of effect size. In fact, the hypothetical effect size can be defined as a single value according to previous results in the literature or experts indications, or by specifying a distribution of plausible values.

No data privacy is involved.

I could not find a proper category for the package. I would suggest something as "preregistration tools" or more general "statistical tools". Following recent concern about replicability crisis in many scientific fields (e.g., psychology and social sciences), this package aims to increase researchers' awareness about inferential risks and the importance of the design phase of the studies. The package documentation is available at https://claudiozandonella.github.io/PRDA/
