### 2020-10-19 #400 @maelle
:wave: @ClaudioZandonella thanks again for your submission! We've however decided it is out-of-scope as it is too statistical... which means it might be in-scope for ["rOpenSci Statistical Software Peer-Review"](https://ropensci.org/stat-software-review/) to be launched in the first half of 2021, so you might try and submit your package to it then. :slightly_smiling_face:
