### 2020-10-19 #400 @ClaudioZandonella
Hi @maelle, thank you for taking the time to review my presubmission. I understand your decision. I want to thank you and the entire rOpenSci community as your documentation and guidelines were really useful and helped me a lot in developing a better package. I really appreciate your work!
